Set up initial project
Updated README with rules and conventions for this project.

Add MIT license.

Add simple options to compiler to implement one of the rules, which is "use absolute imports".

Remove template code that was generated from create-next-app.

Integrate MUI, a component library with built in features, into the project, which led us to the first common module "themes".